chore: improve button component
