+++ components/Form/Button/Button.interface.ts
@@ -1,0 +1,9 @@
+export type ButtonProps = {
+  type?: 'button' | 'submit';
+  onClick: () => void;
+  isDisabled: boolean;
+  haveAnimation?: boolean;
+  className?: string;
+  children: React.ReactNode;
+  loading?: boolean;
+};
